First stab at breaking up the roll
I'm trying to come up with a way to communicate the dice roll better.  This change essentially breaks up the roll by type-of-dice.  The main idea here was to get the rolls bucketized so that I could do an advantage/disadvantage feature.

Note that the unit tests are not passing with this.  I'm still playing around with the format of the result and don't want to convert all of the tests until things have settled.