Add support for advantage/disadvantage
This adds the core support for advantage or disadvantage on the roll. It only supports d20s, and currently you don't get to see the thrown away roll.  Need to figure out how to best bubble this up.

Unit tests are all passing now, even though I think I'm going to change the format of the result.